nambahhin like dan komen

+++ app/Http/Controllers/LikeController.php
@@ -8,20 +8,24 @@ use Illuminate\Http\Request;
 
 class LikeController extends Controller
 {
-    public function like(String $id){
+    public function like(String $id)
+    {
+        if (auth()->check()) {
             $like = Like::where('photoId', $id)->where('userId', auth()->user()->userId)->first();
             if ($like) {
                 $like->delete();
-                return back();
             } else {
                 $tanggal = Carbon::now()->toDateTimeString();
-            // dd($id);
                 $like = new Like();
                 $like->photoId = $id;
                 $like->userId = auth()->user()->userId;
                 $like->date_like = $tanggal;
                 $like->save();
-                return back();
             }
+            return back();
+        } else {
+            // Redirect user to login page if not authenticated
+            return redirect()->route('login')->with('error', 'Anda harus login untuk memberikan suka.');
         }
+    }
 }
